requirements_m1.txt
tensorflow-macos
tensorflow-metal
numpy
pandas
scikit-learn
matplotlib
seaborn
hickle 
requests